Add setting to disable or rename current project root favorite directory Currently we automatically add a favourite directory entry to the file chooser for the project root. These look like "Shotgun Project Root". Multi-root setups have an entry for each root like this: "Shotgun Project Root (secondary)", etc. This can be quite redundant so some studios. And the name of the favourite is quite lengthy. This can make things cluttered in that pane in the file chooser.
This code adds a setting `current_project_favourite_name` allowing you to rename the menu to anything you like (in multi-root setups, the root name will still be automatically added on to the end of the name). Additionally, if you specify an empty string `''` as the value, this will disable adding the menu to the favorites pane.