#36608: Changes a log_warning to a log_debug to reduce noisy output.
This is going to be triggered on most launches of Hiero and Nuke Studio, as both will cause a file read of a template file that is unlikely to be stored in an SGTK-aware location. We don't need to be scaring the user with nasty output containing Python tracebacks when it's completely harmless.