[minor] Imports QtGui locally in menu_generation.py.
This is because we have situations where tk-nuke's Python module is imported very early in the launch process -- specificially on a File->New or File->Open operation. Because of that, sgtk.platform.qt isn't properly initialized yet and we're in a situation where we don't have access to sgtk.platform.qt.QtGui when the import is attempted. We can still use sgtk.platform.qt, but we need to import QtGui from it when it's used and not when the module is imported.