For #46098: Flame OpenClip updating when no templates exist. (#48)
When no template exists to identify an OpenClip file for the current Shot, we attempt to find a published clip file that comes from the new-style Flame workflow that makes use of the publish2 app during sequence export. If a usable clip is found, we update its contents in the same way we've always done in template-driven workflows.