For SG-4782: Skips importing the web engine widgets when running in Nuke 11+ on Windows. (#50)
Importing QtWebEngineWidgets from PySide2 on Windows can cause an unrecoverable deadlock in some cases. Since we do not make use of it in any of the out-of-the-box Toolkit bundles or tk-core, we can safely skip importing that submodule. Any clients that require its use can import it directly from PySide2.